createCRUDV7

+++ hairsalon/src/main/java/com/hairsalonbookingapp/hairsalon/repository/DiscountProgramRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DiscountProgramRepository extends JpaRepository<DiscountProgram, Integer> {
-    DiscountProgram findDiscountProgramById(int id);
+public interface DiscountProgramRepository extends JpaRepository<DiscountProgram, Long> {
+    DiscountProgram findDiscountProgramById(long id);
 }
